moved validateFile to filehelper and renamed it to isFile
